Move certtuil file to pkg directory. (#3611)
